# Speed up WritableBuffer.WriteXxx