--- a/src/Microsoft.AspNetCore.Server.Kestrel/Internal/Http/PipelineExtensions.cs
+++ b/src/Microsoft.AspNetCore.Server.Kestrel/Internal/Http/PipelineExtensions.cs
@@ -194,7 +194,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Internal.Http
 
             var length = _maxULongByteLength - position;
             buffer.Write(new ReadOnlySpan<byte>(byteBuffer, position, length));
-            buffer.Advance(length);
         }
 
         [MethodImpl(MethodImplOptions.NoInlining)]