dev
```
 Method |                       Type |      Mean |    StdDev |       Op/s | Allocated |
------- |--------------------------- |---------- |---------- |----------- |---------- |
 Output |                 LiveAspNet | 3.8515 us | 0.0253 us | 259,636.96 |     664 B |
 Output |           PlaintextChunked | 1.4079 us | 0.0277 us | 710,277.44 |     232 B |
 Output | PlaintextChunkedWithCookie | 2.9045 us | 0.0464 us | 344,288.21 |     814 B |
 Output |        PlaintextWithCookie | 3.4845 us | 0.0525 us | 286,983.30 |     900 B |
 Output |       TechEmpowerPlaintext | 1.8344 us | 0.0136 us | 545,149.42 |     317 B |
```

PR
```
 Method |                       Type |      Mean |    StdDev |       Op/s | Allocated |
------- |--------------------------- |---------- |---------- |----------- |---------- |
 Output |                 LiveAspNet | 3.0829 us | 0.0762 us | 324,367.00 |     353 B |
 Output |           PlaintextChunked | 1.2563 us | 0.0075 us | 795,959.04 |     192 B |
 Output | PlaintextChunkedWithCookie | 1.9574 us | 0.0136 us | 510,893.40 |     487 B |
 Output |        PlaintextWithCookie | 2.6213 us | 0.0294 us | 381,491.96 |     573 B |
 Output |       TechEmpowerPlaintext | 1.7388 us | 0.0293 us | 575,102.39 |     277 B |
```

/cc @davidfowl