@benaadams 2017-03-18
Added benchmarks