@benaadams 2017-03-18
feature-si for reference - still long way to go
```
 Method |                       Type |          Mean |     StdDev |          RPS | Allocated |
------- |--------------------------- |-------------- |----------- |------------- |---------- |
 Output |                 LiveAspNet | 1,384.3867 ns | 21.0264 ns |   722,341.54 |      64 B |
 Output |           PlaintextChunked |   570.4469 ns |  8.7973 ns | 1,753,011.45 |      64 B |
 Output | PlaintextChunkedWithCookie |   927.3479 ns | 10.3280 ns | 1,078,343.95 |      64 B |
 Output |        PlaintextWithCookie |   953.3636 ns | 12.3276 ns | 1,048,917.75 |      64 B |
 Output |       TechEmpowerPlaintext |   581.0874 ns |  7.2057 ns | 1,720,911.39 |      64 B |
```