@benaadams 2017-03-18
@davidfowl Benchmark isn't returning blocks, when/how is that supposed to happen (i.e. what hook do I need to add)