@benaadams 2017-03-18
PR
```
 Method |                       Type |      Mean |    StdDev |      Op/s | Allocated |
------- |--------------------------- |---------- |---------- |---------- |---------- |
 Output |                 LiveAspNet | 2.8919 us | 0.0866 us | 345797.29 |     353 B |
 Output |           PlaintextChunked | 1.2609 us | 0.0103 us | 793106.22 |     192 B |
 Output | PlaintextChunkedWithCookie | 1.9944 us | 0.0231 us | 501415.48 |     487 B |
 Output |        PlaintextWithCookie | 2.2267 us | 0.0186 us | 449097.08 |     509 B |
 Output |       TechEmpowerPlaintext | 1.3983 us | 0.0168 us | 715156.59 |     214 B |
```